ToDotNot unit test conversion - part 10.
