+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesArrayOfArraysOfObjectType/Schema.json
@@ -1,0 +1,19 @@
+{
+  "type": "object",
+  "properties": {
+    "arrayOfArrayOfObject": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/itemType"
+      }
+    }
+  },
+  "definitions": {
+    "itemType": {
+      "type": "array",
+      "items": {
+        "type": "object"
+      }
+    }
+  }
+}
